merged Council game logic into main repos and fixed bug in NavBar
Before, when on the MyPieces page and load a piece, remains on MyPieces page and renders CreatePiece component. When this happens, NavBar would not go pack to original MyPieces page when icon clicked.  Now redirects to the CreatePiece page, fixing bug. Also the Council game logic has been implemented as is currently being tested.